requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiomysql==0.2.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asyncmy==0.2.10
attrs==25.3.0
certifi==2025.8.3
click==8.2.1
databases==0.9.0
fastapi==0.116.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.3
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
python-environ==0.4.54
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
yarl==1.20.1
